More IO protocol, for reading
This implements more of the IO protocol in io.go, with enough now to
do a read (which expects to get a promise).

Note that now we have things happening concurrently with the script
execution, so we have to wait for all I/O to be completed -- including
anything that is added as the result of another I/O completing.